Split roll module
